Windows-related fixes.
We fixed a few compiler-related nitpicks, especially related to printf
format specifiers.

`nqiv_agetc` has been rewritten to `nqiv_stdin_agetc` for the time-being
and functions that call it specifically check whether they are reading
stdin to use it, since in Windows, stdin can be read as a special
console type in addition to a file, depending on context. stdin is so
far, the only thing we need to read asynchronously, but this may change,
later.

`nqiv_stdin_agetc` is also guaranteed to return a 32-bit integer since
we may need a (little endian) UTF-8 byte-sequence encoded therein, as
reading the console may return wide characters which we convert.

Note when reading from the Windows console, we use characters from
key-up events. This might create a few awkward cases, but at this time,
timing key-down events for repetitions is considered an overly-niche
case.

+++ src/thumbnail.c
@@ -263,7 +263,9 @@ bool nqiv_thumbnail_create(nqiv_image* image)
 	char mtime_string[NQIV_MTIME_STRLEN + 1] = {0};
 	snprintf(mtime_string, NQIV_MTIME_STRLEN, "%" PRIuMAX, (uintmax_t)stat_data.mtime);
 	char size_string[NQIV_SIZE_STRLEN + 1] = {0};
-	snprintf(size_string, NQIV_SIZE_STRLEN, "%zu", stat_data.size);
+	/* TODO We cast to uintmax_t to workaround MinGW not handling %zu. Hopefully some day we can avoid this. */
+	/*snprintf(size_string, NQIV_SIZE_STRLEN, "%zu", stat_data.size);*/
+	snprintf(size_string, NQIV_SIZE_STRLEN, "%" PRIuMAX, (uintmax_t)stat_data.size);
 	char width_string[NQIV_DIMENSION_STRLEN + 1] = {0};
 	snprintf(width_string, NQIV_DIMENSION_STRLEN, "%d", image->image.width);
 	char height_string[NQIV_DIMENSION_STRLEN + 1] = {0};
